Changes the approach to defining annual mileage and payload
Change in payload and annual mileage is now function of size class (as before), but also the driving cycle (use). Annual mileage and payload are lower when the vehicle is sued for urban delivery, than when used for long haul. Such change factors are based on https://eur-lex.europa.eu/eli/reg/2019/1242/oj#d1e32-239-1